solving the N issue.
When a context was in a masked region then the position had a N as a biset and the biset_mask function crashed. This is normal but it made me change plenty of things.

See issue #1 in gitea  for more details.

Basically if N is in the context (then the position soud be removed) I generate a all False mask and then a all -1 pvector than then when generates a -1 position in the output table

UNTESTED